Fibonacci.jl

Documentation for Fibonacci.jl.

Fibonacci sequence

Fibonacci sequence is a well-known sequence named after italian mathematician Leonardo da Pisa, also known as Fibonacci.

Starting with $F_1=F_2=1$, each following number $F_n$ is determined by the sum of the previous two numbers $F_{n-1}$ and $F_{n-2}$.

\[F_n = F_{n-1} + F_{n-2}\]

API

Fibonacci.fibonacci

Fibonacci.fibonacci — Function

fibonacci(n)

Return the 'n'-th and 'n+1'-st number of the Fibonacci sequence.

Examples

julia> fibonacci(2)
(1, 2)